Write out per-core particle counts to checkpoints
To resume from a checkpoint, all particles were being read by all
cores, and discarded if they're not owned by a given core. In part,
this may be due to flredecomp not being particle-aware, and therefore
not re-decomposing them.

Here, we add a file attribute to particle checkpoints specifying the
per-core particle count, so that only relevant particles are being
read in. A (failing) test has been added to ensure that when
flredecomp is made aware of checkpointed paricles, they are decomposed
correctly.