Attempt to reproduce detector behavior with particles
Update lagrangian_detectors test, prior to the culling of
detectors. However, at present, results can be replicated for single
particles, but not groups of particles > 1. The order of particles in
the pos array seems different to the order or detectors. This must
either be a feature of Chris' implementation or the new particle IO.