Save particle IDs in dumps and checkpoints
The older IO format would save particles in ID-order, so their order
was preserved regardless of the decomposition. With H5hut, the
particles are written to file batched by processor, which means that
the order is no longer preserved. We save the (as of now)
globally-unique particle IDs to dumps and checkpoints so that
individual particles can be tracked.