Remove allgather in particle checkpointing
Particle checkpointing needs to know the number of particles per
process, so that particles can be read back to the correct
location. This is handled with an HDF5 attribute, which counts as
metadata, and should be written collectively by all processes. This is
because the write does not go through immediately -- instead all
processes maintain a metadata cache. However, the npoints attribute
does not need to be read as part of the checkpointing procedure, so it
is okay if it is incorrect on non-root processes (which don't flush
their metadata cache to disk). As a result, we're able to eliminate
the MPI_Allgather operation, reducing it to a simple MPI_Gather.