wip! start refactoring reading particle attributes
There's a lot of duplication when it comes to figuring out the names
of particle attributes, old attributes, and fields to be stored on
particles. This kind of information doesn't change during the course
of a run, so we should just store it at the start.

This has two useful effects:
1. Big reduction in code duplication, which is always a chance for a
   bug to creep in
2. Clarifies the intent when it comes to particle routines: you can
   see directly how they're manipulating the data, instead of trying
   to parse it out from the surrounding code.