Refactor particle IO, adding vector and tensor attributes
This is a big squash of a couple of unwieldy branches that I forked
off cfa68052c ("Document time_period and reenable CPU and wall time
dumps", 2019-08-06), as part of the particle-io branch. This was
squashed from 7ed4bae3b ("Change suffix on vector/tensor attributes",
2019-10-10).

The main aim of this is to support vector and tensor-valued attributes
on particles, in addition to scalar attributes as were already
implemented. As a part of this, I've done a significant amount of
refactoring to remove code duplication related to querying the
attributes present on particles, etc. (which would have grown
significantly by introducing further types of attributes).