Make output optional on particle groups
This adds a new configuration key,
"particle_group/particle_io/exclude_from_output". If present, the
corresponding particle group, and all its subgroups will not be dumped
to files. This would be useful for maintaining some attributes for
online computations, that do not need to be saved after the fact.