Drop nparticles from set_particle_attribute calls
Because we're still in the Fortran interface at this point, we can
just query the output array size instead of requiring another explicit
argument. Note that we keep the natt argument, because we're
essentially flattening a 3D array into the 2D output array, so it's
clearer, even though we could calculate the number of attributes from
the geometric dimension and array size.