Make from_flredecomp flag in initialise_particles explicit
Instead of just changing behaviour based on the flag, we split it into
two optional parameters that were affected by the flag: whether or not
to setup output files for particles, and whether to ignore analytical
particles, i.e. those which aren't from a file and don't need to be
handled by flredecomp.